Add Scene 6-4 script
This was chosen first because Baby Face would have the most complex monitoring
logic out of all bosses by far, because of both its phase count, the weird
shit it does with its stats in-between phases, and the fact the boss can end
early during its fourth phase if you get caught. This complexity allowed for
more rigourous testing of the script capabilities and helped in catching a bunch
of problems with older implementations of certain script functions that needed
adjusting. Because of this, implementing bosses for other levels should hopefully
be easier and less error-prone.